pyproject.toml
[project]
name = "home-assistant-bme280"
version = "0.0.0-dev"
description = "Simple, auto-discoverable bme280 for Home Assistant"
authors = [
    {name = "swnf"}
]
license = "MIT"
license-files = ["LICENSE.md"]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "ha-mqtt-discoverable (>=0.23.0,<0.24.0)",
    "rpi-bme280 (>=0.2.4,<0.3.0)",
    "smbus2 (>=0.6.0,<0.7.0)"
]

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"